Fixed "error: invalid operands to binary +" errors in rendertarget and shadowprojection samples.

+++ pspsdk/sdk/samples/gu/rendertarget/rendertarget.c
@@ -272,7 +272,7 @@ int main(int argc, char* argv[])
 	Texture offscreenTexture = {
 		GU_PSM_4444,
 		0, 128, 128, 128,
-		sceGeEdramGetAddr() + renderTarget
+		sceGeEdramGetAddr() + (int)renderTarget
 	};
 
 	for(;;)
